better conductor contract pieces
