+++ purchasing_test/integration/scout/test_scout_search.py
@@ -2,8 +2,6 @@
 
 import datetime
 
-from purchasing.database import Model, RefreshSearchViewMixin, Column
-
 from purchasing.app import db
 from purchasing_test.test_base import BaseTestCase
 from purchasing_test.util import (
@@ -117,51 +115,3 @@ class TestScoutSearch(BaseTestCase):
         # make sure that contract types are properly handled
         self.assert200(self.client.get('/scout/search?archived=y&contract_type={}&q='.format(self.contract_type2.id)))
         self.assertEquals(len(self.get_context_variable('results')), 1)
-
-class FakeModel(RefreshSearchViewMixin, Model):
-    __tablename__ = 'fakefake'
-    __table_args__ = {'extend_existing': True}
-
-    id = Column(db.Integer, primary_key=True)
-    description = Column(db.String(255))
-
-    def __init__(self, *args, **kwargs):
-        super(FakeModel, self).__init__(*args, **kwargs)
-
-    @classmethod
-    def record_called(cls):
-        cls.called = True
-
-    @classmethod
-    def reset_called(cls):
-        cls.called = False
-
-    @classmethod
-    def event_handler(cls, *args, **kwargs):
-        return cls.record_called()
-
-class TestEventHandler(BaseTestCase):
-    def setUp(self):
-        super(TestEventHandler, self).setUp()
-        FakeModel.reset_called()
-
-    def test_init(self):
-        self.assertFalse(FakeModel.called)
-
-    def test_create(self):
-        FakeModel.create(description='abcd')
-        self.assertTrue(FakeModel.called)
-
-    def test_update(self):
-        fake_model = FakeModel.create(description='abcd')
-        FakeModel.reset_called()
-        self.assertFalse(FakeModel.called)
-        fake_model.update(description='efgh')
-        self.assertTrue(FakeModel.called)
-
-    def test_delete(self):
-        fake_model = FakeModel.create(description='abcd')
-        FakeModel.reset_called()
-        self.assertFalse(FakeModel.called)
-        fake_model.delete()
-        self.assertTrue(FakeModel.called)
